fix: restructure routing to manual mode and fix Invariant failed errors
- Extract router structure to separate files (router/routeTree.ts, router/index.ts)
- Fix child route paths by removing leading `/` (course/$courseId vs /course/$courseId)
- Replace deprecated getParent with getParentRoute
- Add HMR-safe router singleton to prevent double mounting
- Extract page components from main.tsx to separate files
- Add RouteErrorBoundary with Sentry integration
- Wrap deep-link processing in try-catch for robustness
- Implement proper Outlet usage in root route

Fixes: Invariant failed error in routing-*.js chunk
Addresses: file-based/manual route conflicts, path uniqueness issues
